🔨 [projects] Inline function `linkproject`
